chore: adjust project file naming conventions
